seeded data for user and address

+++ models/user.js
@@ -25,7 +25,7 @@ User.init(
       type: DataTypes.STRING
     },
     phone_number: {
-      type: DataTypes.INTEGER
+      type: DataTypes.BIGINT
     },
   },
   {
